requirements.txt
absl-py==0.9.0
ase==3.19.1
astor==0.8.1
autograd==1.3
autograd-gamma==0.4.2
backcall==0.2.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
ConfigSpace==0.4.12
cycler==0.10.0
Cython==0.27.3
decorator==4.4.2
emcee==3.0.2
future==0.18.2
gast==0.3.3
george==0.3.1
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.29.0
h5py==2.10.0
hpbandster==0.7.4
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
importlib-metadata==1.6.0
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.17.1
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lifelines==0.24.13
lightgbm==2.3.1
llvmlite==0.32.1
Markdown==3.2.2
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.2
-e git+https://github.com/google-research/nasbench@0c898439d48eba0ed01700b7d99fbf101667d0db#egg=nasbench
netifaces==0.10.9
networkx==2.4
ngboost==0.2.0
nose==1.3.7
numba==0.49.1
numpy==1.18.1
numpy-indexed==0.3.5
olefile==0.46
pandas==1.0.3
parso==0.7.0
pathvalidate==2.3.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plyfile==0.7.2
prompt-toolkit==3.0.5
protobuf==3.12.0
psutil==5.7.0
ptyprocess==0.6.0
pybind11==2.5.0
pybnn==0.0.5
Pygments==2.6.1
pyparsing==2.4.7
pyrfr==0.8.0
Pyro4==4.80
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
rdflib==5.0.0
requests==2.23.0
RoBO==0.3.1
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
serpent==1.30.2
six==1.14.0
statsmodels==0.11.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
threadpoolctl==2.0.0
tifffile==2020.5.11
torch==1.5.0
torch-cluster==1.5.4
torch-geometric==1.5.0
torch-scatter==2.0.4
torch-sparse==0.6.3
torch-spline-conv==1.2.0
torchvision==0.6.0a0+82fd1c8
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==1.1.0
zipp==3.1.0
